requirements.txt
coverage==7.4.0  # for testing
django==4.2.9
gunicorn==21.2.0
psycopg2-binary==2.9.9
pytest-django==4.7.0  # for testing
pytest==7.4.4  # for testing
wait-for-it==2.2.2
django-allauth==0.61.1  # for authentication
